requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
ddgs>=9.0.0
python-whois==0.9.6
colorama==0.4.6
python-dotenv==1.2.1
playwright==1.58.0
playwright-stealth==1.0.6
dnspython==2.7.0
tldextract==5.1.3
tiktoken==0.8.0
pillow==10.3.0
fastapi>=0.111.0
uvicorn>=0.30.1
diskcache>=5.6.3
